feat: change logic delete
